Refactoring shader controls to separate folder
