+++ src/Drizzle.UI.UWP/App.xaml
@@ -10,7 +10,7 @@
         <ResourceDictionary>
             <ResourceDictionary.MergedDictionaries>
                 <XamlControlsResources xmlns="using:Microsoft.UI.Xaml.Controls" />
-                <ResourceDictionary Source="UserControls/D2D1AnimatedPixelShaderPanel.xaml" />
+                <ResourceDictionary Source="UserControls/Shaders/D2D1AnimatedPixelShaderPanel.xaml" />
             </ResourceDictionary.MergedDictionaries>
 
             <StaticResource x:Key="WindowCaptionBackground" ResourceKey="ApplicationPageBackgroundThemeBrush" />
