+++ src/Drizzle.UI.UWP/UserControls/Shaders/AnimatedComputeShaderPanelEx.xaml
@@ -1,5 +1,5 @@
 ﻿<UserControl
-    x:Class="Drizzle.UI.UWP.UserControls.AnimatedComputeShaderPanelEx"
+    x:Class="Drizzle.UI.UWP.UserControls.Shaders.AnimatedComputeShaderPanelEx"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:animation="using:CommunityToolkit.WinUI.Animations"
@@ -7,7 +7,7 @@
     xmlns:customControls="using:Drizzle.UI.UWP.UserControls"
     xmlns:customConverters="using:Drizzle.UI.UWP.Converters"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:local="using:Drizzle.UI.UWP.UserControls"
+    xmlns:local="using:Drizzle.UI.UWP.UserControls.Shaders"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     d:DesignHeight="300"
     d:DesignWidth="400"
