+++ src/Drizzle.UI.UWP/UserControls/Shaders/D2D1AnimatedPixelShaderPanel.xaml
@@ -1,7 +1,7 @@
 <ResourceDictionary
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:Drizzle.UI.UWP.UserControls"
+    xmlns:local="using:Drizzle.UI.UWP.UserControls.Shaders"
     xmlns:win2d="using:Microsoft.Graphics.Canvas.UI.Xaml">
     <Style TargetType="local:D2D1AnimatedPixelShaderPanel">
         <Setter Property="Template">
